Check back RegisterActivity
