--- .idea/shelf/Uncommitted_changes_before_Checkout_at_11_17_2020_5_14_PM__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_11_17_2020_5_14_PM_[Default_Changelist]" date="1605608067416" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_11_17_2020_5_14_PM_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 11/17/2020 5:14 PM [Default Changelist]" />
-</changelist>
